+++ crates/rtlola-frontend/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `is_eval_filtered` method to stream trait.
 - Added `memory_bound` method to window trait.
 - Added support for multiple eval clauses.
+- Added methods to export and import the MIR to/from a json representation.
 
 ### Fixed
 - Remap stream references during lowering to MIR to keep streams accessible by index.